# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string",
						"description": "가입된 이메일"
					},
					"name": {
						"type": "string",
						"description": "가입된 유저 이름"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "선택한 선호 카테고리 이름 배열 (예: [\"한식\", \"중식\"])"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorDetail": {
				"properties": {
					"errorCode": {
						"type": "string",
						"description": "에러 코드 (예: U001, M001)"
					},
					"reason": {
						"type": "string",
						"description": "에러 메시지"
					},
					"data": {
						"nullable": true
					}
				},
				"required": [
					"errorCode",
					"reason",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"description": "공통 실패 응답 형식.\ntsoa의",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false
					},
					"error": {
						"$ref": "#/components/schemas/ErrorDetail"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일 (로그인 시 사용)"
					},
					"password": {
						"type": "string",
						"description": "비밀번호 (서버에서 bcrypt로 해싱하여 저장)"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"gender": {
						"type": "string",
						"description": "성별 (예: \"male\", \"female\")"
					},
					"birth": {
						"type": "string",
						"format": "date-time",
						"description": "생년월일 (예: \"1999-01-01\")"
					},
					"address": {
						"type": "string",
						"description": "기본 주소 (선택)"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소 (선택)"
					},
					"phoneNumber": {
						"type": "string",
						"description": "휴대폰 번호 (예: \"010-1234-5678\")"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 카테고리 ID 배열 (예: [1, 2])"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"email": {
						"type": "string",
						"description": "유저 이메일"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					}
				},
				"required": [
					"id",
					"email",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserUpdateResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserUpdateResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"gender": {
						"type": "string",
						"description": "성별 (예: \"male\", \"female\")"
					},
					"birth": {
						"type": "string",
						"format": "date-time",
						"description": "생년월일 (예: \"1999-01-01\")"
					},
					"address": {
						"type": "string",
						"description": "기본 주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "휴대폰 번호 (예: \"010-1234-5678\")"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReviewStore": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"name": {
						"type": "string",
						"nullable": true,
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"nullable": true,
						"description": "가게 주소"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID"
					}
				},
				"required": [
					"id",
					"name",
					"address",
					"regionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewUser": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"email": {
						"type": "string",
						"description": "유저 이메일"
					}
				},
				"required": [
					"id",
					"name",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"body": {
						"type": "string",
						"description": "리뷰 본문"
					},
					"score": {
						"type": "number",
						"format": "double",
						"description": "별점 (1~5)"
					},
					"store": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReviewStore"
							}
						],
						"nullable": true,
						"description": "리뷰가 달린 가게 정보"
					},
					"user": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReviewUser"
							}
						],
						"nullable": true,
						"description": "리뷰를 작성한 유저 정보"
					}
				},
				"required": [
					"id",
					"body",
					"score",
					"store",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"description": "리뷰 목록 조회 응답",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewItem"
						},
						"type": "array",
						"description": "리뷰 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 커서 (다음 페이지 없으면 null)"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddReviewResponse": {
				"description": "리뷰 작성 성공 응답.\ncreateReview 서비스가 실제로 반환하는 객체 구조를 타입으로 명시한 것.",
				"properties": {
					"reviewId": {
						"type": "number",
						"format": "double",
						"description": "생성된 리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰가 등록된 가게 ID"
					}
				},
				"required": [
					"reviewId",
					"storeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AddReviewResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/AddReviewResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddReviewRequest": {
				"properties": {
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰를 작성할 가게 ID"
					},
					"body": {
						"type": "string",
						"description": "리뷰 본문"
					},
					"score": {
						"type": "number",
						"format": "double",
						"description": "별점 (1~5)"
					}
				},
				"required": [
					"storeId",
					"body",
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddMissionResponse": {
				"description": "미션 추가 성공 응답.\ncreateMission 서비스가 실제로 반환하는 객체 구조를 타입으로 명시한 것.\ntsoa",
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "생성된 미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "미션이 등록된 가게 ID"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "미션 완료 시 지급할 포인트"
					},
					"deadline": {
						"type": "string",
						"format": "date-time",
						"description": "미션 마감일"
					},
					"missionSpec": {
						"type": "string",
						"description": "미션 내용 설명"
					}
				},
				"required": [
					"missionId",
					"storeId",
					"reward",
					"deadline",
					"missionSpec"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AddMissionResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/AddMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddMissionRequest": {
				"properties": {
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "미션을 등록할 가게 ID"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "미션 완료 시 지급할 포인트"
					},
					"deadline": {
						"type": "string",
						"description": "미션 마감일 (예: \"2025-12-31\")"
					},
					"missionSpec": {
						"type": "string",
						"description": "미션 내용 설명"
					}
				},
				"required": [
					"storeId",
					"reward",
					"deadline",
					"missionSpec"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionStore": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"name": {
						"type": "string",
						"nullable": true,
						"description": "가게 이름"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"missionSpec": {
						"type": "string",
						"description": "미션 내용 설명"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "미션 완료 시 지급할 포인트"
					},
					"deadline": {
						"type": "string",
						"format": "date-time",
						"description": "미션 마감일"
					}
				},
				"required": [
					"id",
					"missionSpec",
					"reward",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListByStoreResponse": {
				"description": "특정 가게 기준 미션 목록 조회 응답",
				"properties": {
					"store": {
						"$ref": "#/components/schemas/MissionStore",
						"description": "조회된 가게 정보"
					},
					"missions": {
						"items": {
							"$ref": "#/components/schemas/MissionItem"
						},
						"type": "array",
						"description": "해당 가게의 미션 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 커서 (다음 페이지 없으면 null)"
					}
				},
				"required": [
					"store",
					"missions",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChallengeMissionResponse": {
				"description": "미션 도전 성공 응답.\nchallengeMission 서비스가 실제로 반환하는 객체 구조를 타입으로 명시한 것.",
				"properties": {
					"userMissionId": {
						"type": "number",
						"format": "double",
						"description": "생성된 유저-미션 관계 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "도전하는 유저 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "도전한 미션 ID"
					},
					"status": {
						"type": "string",
						"description": "미션 진행 상태 (예: \"진행중\")"
					}
				},
				"required": [
					"userMissionId",
					"userId",
					"missionId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChallengeMissionResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ChallengeMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChallengeMissionRequest": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "도전할 미션 ID"
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserInfo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListByUserResponse": {
				"description": "사용자 기준 도전 중인 미션 목록 조회 응답",
				"properties": {
					"user": {
						"$ref": "#/components/schemas/UserInfo",
						"description": "조회된 유저 정보"
					},
					"missions": {
						"items": {
							"$ref": "#/components/schemas/MissionItem"
						},
						"type": "array",
						"description": "도전 중인 미션 목록"
					},
					"stores": {
						"items": {
							"$ref": "#/components/schemas/MissionStore"
						},
						"type": "array",
						"description": "미션과 1:1 매칭되는 가게 정보 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 커서 (다음 페이지 없으면 null)"
					}
				},
				"required": [
					"user",
					"missions",
					"stores",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"409": {
						"description": "이미 존재하는 이메일",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "이메일, 비밀번호, 선호 카테고리 등을 입력받아 회원가입을 처리합니다.",
				"summary": "회원가입",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/me": {
			"patch": {
				"operationId": "HandleUpdateUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserUpdateResponse_"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "로그인한 사용자의 프로필 정보를 수정합니다.\nGoogle 로그인으로 가입한 경우 비어있는 생일, 전화번호 등을 채울 때 사용합니다.",
				"summary": "내 프로필 수정",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/reviews": {
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewListResponse"
								}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게의 리뷰 목록을 cursor 기반 페이지네이션으로 조회합니다.",
				"summary": "가게 리뷰 목록 조회",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores/reviews": {
			"post": {
				"operationId": "HandleAddReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AddReviewResponse_"
								}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게에 리뷰를 작성합니다. (로그인 필요)",
				"summary": "리뷰 작성",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddReviewRequest"
							}
						}
					}
				}
			}
		},
		"/me/reviews": {
			"get": {
				"operationId": "HandleListMyReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewListResponse"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "내가 작성한 리뷰 목록을 cursor 기반 페이지네이션으로 조회합니다. (로그인 필요)",
				"summary": "내 리뷰 목록 조회",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores/missions": {
			"post": {
				"operationId": "HandleAddMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AddMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게에 새로운 미션을 추가합니다. (로그인 필요)",
				"summary": "가게 미션 추가",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddMissionRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/missions": {
			"get": {
				"operationId": "HandleGetMissionsByStore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionListByStoreResponse"
								}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게의 미션 목록을 cursor 기반 페이지네이션으로 조회합니다.",
				"summary": "가게 미션 목록 조회",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/challenge": {
			"post": {
				"operationId": "HandleChallengeMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChallengeMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "이미 도전 중인 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "가게의 미션을 내 '도전 중인 미션' 목록에 추가합니다. (로그인 필요)",
				"summary": "미션 도전 시작",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChallengeMissionRequest"
							}
						}
					}
				}
			}
		},
		"/me/missions": {
			"get": {
				"operationId": "HandleGetMissionByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionListByUserResponse"
								}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 사용자",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "인증 실패",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "내가 현재 도전 중인 미션 목록을 cursor 기반 페이지네이션으로 조회합니다. (로그인 필요)",
				"summary": "내 도전 중인 미션 목록 조회",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}